+++ sentry-spring-jakarta/src/test/kotlin/io/sentry/spring/jakarta/tracing/SentryCaptureExceptionParameterAdviceTest.kt
@@ -2,6 +2,8 @@ package io.sentry.spring.jakarta.tracing
 
 import io.sentry.IHub
 import io.sentry.exception.ExceptionMechanismException
+import io.sentry.spring.jakarta.exception.SentryCaptureExceptionParameter
+import io.sentry.spring.jakarta.exception.SentryCaptureExceptionParameterConfiguration
 import org.junit.runner.RunWith
 import org.mockito.kotlin.check
 import org.mockito.kotlin.mock
